pyproject.toml
[tool.poetry]
name = "exceltodb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.1"
psycopg2-binary = "^2.9.10"
ipython = "^9.2.0"
django-extensions = "^4.1"
python-dotenv = "^1.1.0"
openpyxl = "^3.1.5"
gunicorn = "^23.0.0"
djangorestframework = "3.14"
django-debug-toolbar = "<5.0.0"
django-mptt = "^0.17.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
